% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_rml.R
\name{dr_rml}
\alias{dr_rml}
\title{Fit the DR-RML model}
\usage{
dr_rml(
  Y,
  Censor,
  A,
  X,
  L,
  PS = c("logit", "logit2", "SL", "GBM"),
  Reg = c("lm", "lm2", "SL"),
  nboot
)
}
\arguments{
\item{Y:}{event time.}

\item{Censor:}{censoring indicator, 1: observed; 0: right-censored.}

\item{A:}{binary treatment indicator 1: treated; 0: untreated.}

\item{X:}{baseline covariate.}

\item{L:}{a scalar value to truncate the timepoint for RML, \code{L} < max(\code{Y}).}

\item{PS:}{specify the propensity score model from one of following options:
\itemize{
  \item \code{logit}: parametric logistic regression.
  \item \code{logit2}: improved parametric logistic regression (Cao et al., 2009).
  \item \code{SL}: super learner (package: \pkg{SuperLearner}) with libraries \code{SL.glm, SL.glm.interaction, SL.step}.
  \item \code{GBM}: Generalized boosted model (package: \pkg{gbm}).
}}

\item{Reg:}{specify the outcome regression model from one of follwoing options:
\itemize{
  \item \code{lm}: parametric gaussian regression.
  \item \code{lm2}: improved parametric gaussian regression (Cao et al., 2009).
  \item \code{SL}: super learner (package: \pkg{SuperLearner}) with libraries \code{SL.glm, SL.glm.interaction, SL.step}.
}}

\item{nboot:}{a numeric value to specify the number of bootstrap for standard error of ACE. If standard error estiamtes is unnecessary, use \code{nboot = 0}.}
}
\value{
\code{dr_RML} returns a list containing at least the following components:
\itemize{
  \item \code{mu1}: average effect for treated group.
  \item \code{mu0}: average effect for untreated group.
  \item \code{ace}: average causal effect \code{mu1} - \code{mu0}.
  \item \code{se}: standard error estimates for \code{ace}.
}
}
\description{
Estimate double-robust average causal effect for restricted mean lifetime.
}
\details{
see Choi et al., (2022+) for detailed method explanation.
}
\examples{
\dontrun{
simdata = function(n, tau) {
  require(mvtnorm)
  expit = function(x)
    exp(x) / (1 + exp(x))
  Sigma = diag(rep(1, 3))
  Sigma[1, 3] = Sigma[3, 1] = 0.2
  Z = rmvnorm(n,
              mean = rep(0, 3),
              sigma = Sigma,
              method = "chol")
  Z1 = Z[, 1]
  Z2 = Z[, 2]
  Z3 = Z[, 3]
  # trt indicator (Bernoulli)
  A = rbinom(n, 1, expit(-0.5 * Z1 - 0.5 * Z2))# Z1 and Z2 are involved in trt assignment
  par.t = exp((-2.5 - 1.5 * Z1 - 1 * Z2 - 0.7 * Z3) * (A == 0) +
                (-3 - 1 * Z1 - 0.9 * Z2 - 1 * Z3) * (A == 1))
  T = rexp(n, par.t)                   # true surv time
  C = rexp(n, tau)                     # independent censoring
  # tau as a controller of censoring rate
  Y = pmin(T, C)                       # observed time
  delta = ifelse(T <= C, 1, 0)         # censoring indicator
  simdata = data.frame(Y = round(Y, 5), delta, A, Z1, Z2, Z3, T)
  simdata[order(Y), ]
}
L = 10       
n = 600      
tau = exp(-4.5)
truepar = ifelse(L == 10, 0.871, 1.682)

set.seed(123)
dt = simdata(n, tau)
Y = dt$Y
Censor = dt$delta
A = dt$A

X = dt[, 4:6]
dr_rml(Y = Y, Censor = Censor, A = A, X = X, L = L, 
       PS = "logit", Reg = "lm", nboot = 10)$ace

# Data Application
data(gse)
Y = gse$Y
Censor = gse$Censor
A = gse$trt
X = gse[, 3:6]
L = 365 * 5
dr_rml(Y = Y, Censor = Censor, A = A, X = X, L = L, 
       PS = "logit", Reg = "lm", nboot = 10)

library(tmle)
library(pseudo)
prmst = pseudomean(Y, Censor, L)
Xt = cova = as.data.frame(model.matrix( ~ -1 + Age + Size + Grade + Er, data = gse))
fit = tmle(
  Y = prmst, A = A, W = cova,
  Q.SL.library = c("SL.glm", "SL.glm.interaction", "SL.step"),
  g.SL.library =  c("SL.glm", "SL.glm.interaction", "SL.step")
)
# Other methods
tmsl = fit$estimates$ATE$psi
library(grf)
cffit = causal_forest(Xt, prmst, A)
cf = mean(cffit$predictions)
}
}
\references{
Cao, W., Tsiatis, A. A., & Davidian, M. (2009). Improving efficiency and robustness of the doubly robust estimator for a population mean with incomplete data. \emph{Biometrika}, \bold{96}(3), 723--734.

Choi, S., Choi, T., Lee, H. Y., Han, S. W., Bandyopadhyay, D. (2022+). Double-robust methods for estimating differences in restricted mean lifetimes using pseudo-observations. \emph{In revision}.
}
