% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gse}
\alias{gse}
\title{GSE6532 cohort study dataset (Loi et al., 2007)}
\format{
A data frame including 374 observations with 7 variables:
\describe{
\item{Y}{time-to-distant metastasis free in days.}
\item{Censor}{censoring indicator (1: distant metastasis, 0: censored).}
\item{Age}{age in year.}
\item{Grade}{Elston-Ellis grading system, 0 for non-cancer patients.}
\item{Size}{tumor size in centimeter.}
\item{Er}{estrogen receptor status excluding node (indicator of individual’s nodal status).}
\item{trt}{1: Tamoxifen treated group, 0: untreated group.}
}
}
\source{
\url{www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=gse6532}
}
\usage{
gse
}
\description{
GSE6532 cohort study data of tamoxifen treatment for breast cancer patients to compare average causal effect in terms of RML (Loi et al., 2007). 
The dataset consists of 374 breast cancer patients: 255 patients in the tamoxifen treatment  group and 119 patients in the control group.
}
\examples{
\dontrun{
data(gse)
Y = dat$Y
Censor = dat$Censor
A = dat$trt
X = dat[, 3:6]
L = 365 * 5
dr_rml(Y = Y, Censor = Censor, A = A, X = X, L = L, PS = "logit", Reg = "lm", nboot = 10)

library(tmle)
prmst = pseudomean(Y, Censor, L)
Xt = cova = as.data.frame(model.matrix( ~ -1 + Age + Size + Grade + Er, data = dat))
fit = tmle(
  Y = prmst, A = A, W = cova,
  Q.SL.library = c("SL.glm", "SL.glm.interaction", "SL.step"),
  g.SL.library =  c("SL.glm", "SL.glm.interaction", "SL.step")
)
# Other methods
tmsl = fit$estimates$ATE$psi
library(grf)
cffit = causal_forest(Xt, prmst, A)
cf = mean(cffit$predictions)
}
}
\references{
Loi, S., Haibe-Kains, B., Desmedt, C., et al. (2007). Definition of clinically distinct molecular subtypes in estrogen receptor-positive breast carcinomas through genomic grade. \emph{Journal of clinical oncology}, \bold{25}(10), 1239--1246.
}
\keyword{dataset}
